requirements.txt
pytest-playwright
fake-useragent